code/inference/requirements.txt
diffusers>=0.33.1
transformers
numpy
pillow
sentencepiece
opencv-python
imageio
imageio[ffmpeg]
tqdm
